Package the JTS library into docker image
This commit adds JTS library into Docker images for both 5.1 & 6.0 versions. Java Topology
Suite (JTS) is a library for creating and manipulating vector geometry that is used by DSE
Search for handling of polygonal shapes.

Before version 1.14 this library was distributed under LGPL license and we can't pack it
into our distribution, so customers were need to install it themselves.  Since version
1.15.0 it has new license that allows to distribute it.

To use new library we also need to update library `spatial4j` that uses JTS - the main
change is in the Java package names, everything works with new version of it as well.